Limpeza de código em componentes e page Home
